refactor(alita): 简化main文件
